[BPF] move conntrack.Timeouts to bpf/conntrack/timeouts
To make it usable ni generic code in collector
